refactor: split kiroclient (aws_error/backoff/idle_reader) and auth (credentials_parser) (M5-1..M5-3)
